requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.5
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
packaging==24.1
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
requests==2.32.3
setuptools==75.1.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
